SoundEnded Message
The track now changes by sleeping a thread until the sound has stopped
playing, and then sending the SoundEnded Message through

However, this means that when the next/prev events are triggered, the
SoundEnded event also triggers.

So there is a rx.recv() at the end of the block to catch the SoundEnded
so it is not sent. But it breaks if I put it before the other lines of
code like updating tui. They do not run. I do not know why this happens.